\name{clean1}
\alias{clean1}
\title{Used by \code{cran.net} for data preparation}
\description{
Used by \code{cran.net} for data preparation. Function is not intended to be called seperately.
}
\usage{clean1(x)}
\arguments{
\item{x}{A vector of author names to be cleaned.}
}
\value{
A cleaned vector.
}
\author{Angela Bohn (\url{angela.bohn@gmail.com})}