\name{emon}
\alias{emon}
\docType{data}
\title{The emon Data Set: Interorganizational Search and Rescue Networks}
\description{
    Seven \pkg{network} objects (edgelists) showing interaction frequencies of
    emergent multi-organizational networks (EMONs) in the context of search and
     rescue (SAR) activities.
}
\usage{data(emon,package="network")}
\details{
    See \code{?network::emon} for further details.
}
\source{The \pkg{network} package.}
\examples{
data(emon,package="network")
}
\keyword{datasets}