\name{sbXXIX}
\alias{sbXXIX}
\alias{sbXXIX_authors}
\alias{sbXXIX_keywords}
\alias{sbXXIX_twomode}
\docType{data}
\title{The sbXXIX Data Set}
\description{
    bla.
}
\usage{data(sbXXIX)}
\details{
    This is a data dump of the INSNA Sunbelt XXIX (2009) abstract database
    for all abstracts accepted for Sunbelt XXIX (2009) poster/presentation.
    \code{sbXXIX_authors}
    Network properties:
    Number of nodes: 112 (58 adjectives, 54 nouns),
    undirected, unvalued,
    Number of edges: 850,
    Number of components: 1.
    \code{sbXXIX_words}
    \code{sbXXIX_twomode}
}
\source{
    INSNA (2009). Sunbelt XXIX Abstract Data. \url{http://www.insna.org/software/data.html}
}
\keyword{datasets}