\name{coleman}
\alias{coleman}
\docType{data}
\title{The coleman Data Set: Coleman's High School Friendship Data}
\description{
    James Coleman (1964) reports research on self-reported friendship
     ties among 73 boys in a small high school in Illinois over the
     1957-1958 academic year.
}
\usage{data(coleman,package="sna")}
\details{
    Two adjacency matrices. See \code{?sna::coleman} for further details.

    Network properties:
    Number of nodes: 73 and 73,
    directed, unvalued,
    Number of edges: 243 and 263,
    Number of components: 5 and 5.
}
\source{The \pkg{sna} package.}
\examples{
data(coleman,package="sna")
}
\keyword{datasets}