\name{emon}
\alias{emon}
\docType{data}
\title{Interorganizational networks (edgelist)}
\description{
    Seven \pkg{network} objects (edgelists) showing interaction frequencies of
    emergent multi-organizational networks (EMONs) in the context of search and
     rescue (SAR) activities.
}
\usage{data(emon,package="network")}
\format{\pkg{network} (edgelist)}
\details{
    See \code{?network::emon} for further details.
}
\source{The \pkg{network} package.}
\references{Drabek, T.E.; Tamminga, H.L.; Kilijanek, T.S.; and Adams, C.R.
     (1981).  Data from Managing Multiorganizational Emergency
     Responses: Emergent Search and Rescue Networks in Natural Disaster
     and Remote Area Settings. Program on Technology, Environment, and
     Man Monograph 33.  Institute for Behavioral Science, University of
     Colorado.
}
\examples{
data(emon,package="network")
}
\keyword{datasets}